livelens/backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.19
pydantic==2.10.3
pydantic-settings==2.6.1
google-generativeai==0.8.3
google-cloud-firestore==2.19.0
google-cloud-storage==2.19.0
playwright==1.49.1

